+++ packages/mangrove.js/test/util/helpers.ts
@@ -31,28 +31,6 @@ export const asyncQueue = <T>(): AsyncQueue<T> => {
   };
 };
 
-export class Deferred<T = any> {
-  public readonly promise: Promise<T>;
-  #resolve: (value?: T | PromiseLike<T>) => void;
-  #reject: (reason?: any) => void;
-
-  constructor() {
-    this.promise = new Promise<T>((resolve, reject) => {
-      this.#resolve = resolve;
-      this.#reject = reject;
-    });
-  }
-
-  resolve(value?: T | PromiseLike<T>): void {
-    this.#resolve(value);
-  }
-
-  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
-  reject(reason?: any): void {
-    this.#reject(reason);
-  }
-}
-
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
 export const toWei = (v: string | number, u = "ether"): BigNumber =>
   utils.parseUnits(v.toString(), u);